Revert change to FileUsage (f64)
Causing issues with backup. The f64 workaround is handled in Core.

+++ src/common_messages.rs
@@ -359,9 +359,8 @@ pub struct FilehostForInstanceRequest {
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct FileUsage {
-    // Note: using f64 rather than usize/u64 because of random bug loading large values with mongo client 2.8.1
-    pub count: Option<f64>,
-    pub size: Option<f64>,
+    pub count: Option<usize>,
+    pub size: Option<usize>,
 }
 
 #[derive(Serialize, Deserialize)]
